# bln-backend/api/request/entry_orders_request.go
package request

import "time"

type EntryOrdersRequest struct {
	NftListID uint   `json:"nftListId" binding:"required"` // NFTId
	Seller    string `json:"seller" binding:"required"`    // 卖家地址
	TokenId   uint   `json:"tokenId" binding:"required"`   // tokenId
	// Price 价格（BT）
	Price     float64   `json:"price" binding:"required"`     // 价格（wei 字符串）
	Deadline  time.Time `json:"deadline" binding:"required"`  // 截止时间
	Nonce     *int      `json:"nonce" binding:"required,gte=0"` // 非重复 nonce（允许 0）
	Signature string    `json:"signature" binding:"required"` // EIP-712 signature
}